requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
build==1.0.3
CacheControl==0.14.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.1.0
crashtest==0.4.1
cryptography==41.0.4
distlib==0.3.7
dulwich==0.21.6
fastjsonschema==2.18.1
filelock==3.12.2
frozenlist==1.4.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.2.0
lxml==5.2.2
magic-filter==1.0.12
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.5
packaging==23.2
pexpect==4.8.0
pipenv==2023.12.1
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.3
tomli==2.0.1
tomlkit==0.12.1
trove-classifiers==2023.9.19
typing_extensions==4.11.0
urllib3==2.0.6
virtualenv==20.25.1
yarl==1.9.4
zipp==3.17.0
